Replace Sun with sunshine glow as main sun component
